split out slow query model class
